requirements.txt
torch
torchvision
gradio
pyyaml
tqdm
numpy
pillow